backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
feedparser>=6.0.0
pgvector>=0.3.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
